Move functions into util file
